+++ internal/idgen/short.go
@@ -12,14 +12,18 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-package cmd
+package idgen
 
 import (
-	"github.com/cardinalhq/lakerunner/internal/idgen"
+	crand "crypto/rand"
+	"encoding/base32"
+	"strings"
 )
 
-// generateOperationID creates a unique operation ID for tracking atomic operations
-// Format: randomBase32
-func generateOperationID(index int) string {
-	return idgen.GenerateShortBase32ID()
+// GenerateShortBase32ID creates a short random base32 ID for operation tracking.
+// it is 8 characters long, and should not be used for security-sensitive operations.
+func GenerateShortBase32ID() string {
+	b := make([]byte, 5) // 5 bytes = 8 base32 chars
+	_, _ = crand.Read(b) // errors from rand.Read are rare and not critical for operation IDs
+	return strings.ToLower(base32.StdEncoding.EncodeToString(b))
 }
